refactor: use S3 base class to setup a bucket for integration tests
