+++ services/distribution/src/test/java/app/coronawarn/server/services/distribution/objectstore/integration/ObjectStoreAccessIT.java
@@ -12,17 +12,11 @@ import app.coronawarn.server.services.distribution.objectstore.client.ObjectStor
 import app.coronawarn.server.services.distribution.objectstore.client.S3Object;
 import app.coronawarn.server.services.distribution.objectstore.publish.LocalFile;
 import app.coronawarn.server.services.distribution.objectstore.publish.LocalGenericFile;
-import com.amazonaws.auth.AWSCredentials;
-import com.amazonaws.auth.BasicAWSCredentials;
-import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.AmazonS3Client;
-import com.amazonaws.services.s3.S3ClientOptions;
 import java.io.IOException;
 import java.nio.file.Path;
 import java.util.List;
 import java.util.UUID;
 import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
@@ -41,7 +35,7 @@ import org.springframework.test.context.junit.jupiter.SpringExtension;
 @DirtiesContext
 @Tag("s3-integration")
 @ActiveProfiles("integration-test")
-class ObjectStoreAccessIT {
+class ObjectStoreAccessIT extends BaseS3IntegrationTest {
 
   private static final String testRunId = "testing/cwa/" + UUID.randomUUID().toString() + "/";
   private static final String rootTestFolder = "objectstore/";
@@ -56,22 +50,6 @@ class ObjectStoreAccessIT {
   @Autowired
   private ObjectStorePublishingConfig objectStorePublishingConfig;
 
-  @BeforeAll
-  public static void setupBucket() {
-    AWSCredentials credentials = new BasicAWSCredentials("accessKey1",
-        "verySecretKey1");
-
-    // Create a client connection based on credentials
-    AmazonS3 s3client = new AmazonS3Client(credentials);
-    s3client.setEndpoint("http://localhost:8003");
-    s3client.setS3ClientOptions(S3ClientOptions.builder().setPathStyleAccess(true).build());
-
-    // Create bucket
-    String bucketName = "cwa";
-    if (!s3client.doesBucketExistV2(bucketName)) {
-      s3client.createBucket(bucketName);
-    }
-  }
 
   @BeforeEach
   public void setup() {
